[Client] Implement way of setting up FX controllers
Because of FX controller are initialized by the FX framework when FXML file is loaded, it was impossible to perform anything after controller is initialized, since it's not us instantiating controller.

By introducing `onLoaded` method in BaseController and calling then it directly after FXML is loaded (controller is ready then), it is possible to run something on controller directly after it's created.

In order to do that it is necessary to:
 * overwrite `onLoaded` method in the subclass of BaseController (eg. EntryController)
 * put whatever is necessary there (layout modification, setting up some behaviour listeners/callbacks etc.)