[Client] Implement UI<->Service coordination for chat retrieving
The purpose of this commit is to make chat data (users & messages) ready to display in the UI layer (handled currently by ChatboardController).
This is achieved by making use of Observer design pattern. Java contains already implementation of this pattern which is represented in `java.util.Observable` and `java.util.Observer` classes.
Both of mentioned classes were used in this implementation